add dryrun to run as presubmit

+++ perf/benchmark/run_benchmark_job.sh
@@ -41,6 +41,7 @@ export USE_MASON_RESOURCE="${USE_MASON_RESOURCE:-True}"
 export CLEAN_CLUSTERS="${CLEAN_CLUSTERS:-True}"
 export NAMESPACE=${NAMESPACE:-'twopods-istio'}
 export PROMETHEUS_NAMESPACE=${PROMETHEUS_NAMESPACE:-'istio-system'}
+export DRYRUN=${DRYRUN:-False}
 
 function setup_metrics() {
   # shellcheck disable=SC2155
@@ -84,6 +85,9 @@ function get_benchmark_data() {
 }
 
 function exit_handling() {
+  if [[ "${DRYRUN}" == "True" ]]; then
+     return
+  fi
   # copy raw data from fortio client pod
   kubectl --namespace "${NAMESPACE}" cp "${FORTIO_CLIENT_POD}":/var/lib/fortio /tmp/rawdata -c shell
   gsutil -q cp -r /tmp/rawdata "gs://${GCS_BUCKET}/${OUTPUT_DIR}/rawdata"
@@ -208,6 +212,10 @@ for dir in "${CONFIG_DIR}"/*; do
        source prerun.sh
     fi
 
+    # dryrun as safe check for presubmit
+    if [[ "${DRYRUN}" == "True" ]]; then
+       continue
+    fi
     # collect config dump after prerun.sh and before test run, to verify test setup is correct
     collect_config_dump
 
@@ -234,6 +242,10 @@ for dir in "${CONFIG_DIR}"/*; do
 
     popd
 done
+
+if [[ "${DRYRUN}" == "True" ]]; then
+   get_benchmark_data "${WD}/configs/dryrun.yaml"
+fi
 
 #echo "collect flame graph ..."
 #collect_flame_graph
